Move error_code to separate file
